Move restore-expiration processing to transition processor
The conductor checks the lag of the object-processor to detect if the
previous lifecycle listing has been processed, i.e. status has been
persisted for all triggered operations.

Consequently it must check the bucket-processor lag (e.g. is listing
terminated), but also the data-mover topic (for transitions) and object-
processor (for expirations). It may check the cold-archive topics, but
it is not needed since the object is immediately updated with a
"transition in process" flag.

However, it must not check topics/consumers which are not directly
linked to the lifecycle listing: in particular the transition-processor,
which handles the end of transitions (update status, requeue...).

To be consistent and avoid unexpected throtlling, restored object
expiration messages must be ignored by the conductor: hence we move
processing to the transition processor.

Issue: BB-449