Bucket-processor: limit the amount of retries
Limit the total amount of retries when processing a range, to keep the
processing time of the kafka message under check (--> avoid rebalance),
yet keep some resiliance to temporary errors.

It does not really matter that we don't retry a lot, since lfiecycle
will be retried later.

Aslo, propagate the error from _applyTransitionRule() so it may get
retried.

Issue: BB-437