Store offset in BackbeatConsumer bootstrap
The purpose of bootstrap is to ensure the consumer receives next
messages: i.e. that the offset is already stored, so that we get every
later message.

For this to work, the offset needs to be comitted: since however we are
not using the `auto.offset.store`, we must manually store the offset.

Issue: BB-441