change connector partition each time it is created
The partition name is used as an indicator in the offsets topic to
distinguish between resume tokens of the different connectors.

When a connector is respawned it should start processing event
from the current moment and not use the previous resume token as it
can be invalid if the connector stayed shut for a long time and the
mongo oplog rotated.

Issue: BB-463