Ensure we try to fill the queue if partially full
If the first call to `consume()` does not return enough entries, the
consumer will wait until an entry is processed to retry consuming. This
may be an issue and delay processing of entries, esp. if for slower
tasks.

We also now ensure we clear the timer when trying to consume(), to keep
the expected semantics of consuming when we finish a task, else 1 second
after last attempt.

Issue: BB-440