Publish offsets before unassign
This is the only place where we can do it (esp. on shutdown), since it
calls `consumer.position()`, with no param, and thus updates the position
of assigned partitions.

This however has an unfortunate effect: it puts Zookeeper in the path,
as the unassign with be delayed until we have retrieved the offsets
and written them to zookeeper.

This should however be fine in the context of a kafka rebalance, which
takes seconds or more.

Issue: BB-455