Set transition-in-progress flag for Transition as well
It used to be done only for archive. It more critical in that case,
since the conductor will not be throttled by objects pending archival (in
sorbet queues).

In case of "simple" transition, the entry is added to DataMover topic,
which is checked by the conductor to wait for the end of a listing
cycle. But the flag may still be needed to avoid retrying to transition
during the actual replication of the data.

It is more consistent to do it in all cases, may avoid some ops, and
will not hurt anyway. The extra metadata IO is needed anyway, to store
the transition date, needed for lifecycle metrics.

Issue: BB-447