Delete orphaned cold object
When an object is deleted (from S3) while it was archived, a message
is sent to the cold storage backend to notify the removal.

When this is happens during the transition (archival) process, the object
is not yet archived, and we cannot sent such message to the backend. In most
cases, this should be fine, as the archiving will fail with an error related
to the absence of the (removed) object.

However, there is a race condition, and it is possible that the archival succeeds
in backend if the object is removed at the end of this process or in between this
result from the backend and the update of metadata in backbeat : the metadata
update will fail (as expected), but the object will still be in cold storage and
remain there forever (not referenced from anywhere, and thus orphaned)

Issue: BB-469