use mongo operation time as notification timestamp
We use the lastModified field to set the timestamp of the notifications,
however this field is only updated when the object’s data is modified
(meaning only on putObject operations).

Change stream entries contain a field containing the timestamp of the
operation in MongoDB, we will use that instead.

Issue: BB-458