Pause consumer when breakbeat is triggered
The current implementation just skips the call to consume()
when breakbeat is triggered, this will cause the consumer
to be kicked out of the consumer group after the max poll
delay is reached.

Polling is only done when calling consume(), hence we should
keep calling that function and instead use the pause/resume
methods that pause/resume the partition fetchers.

Pre-fetched messages are supposedly not purged from the internal
queue, however they are not returned when calling consume() in a
paused state.

The pause/resume state of the partitions is only kept locally, so
this implementation is safe in case of forced shutdown or restart.

Issue: BB-456