simplify handling of replication failure notification
Instead of duplicating the notification logic we set the
correct originOp when updating the metadata to trigger the
replication failure notification.

At the end of a failed replication we update the metadata
with the failure status that ends up in the oplog topic.
So this does not add any overhead in terms of extra Kafka messages
being pushed to any of the topics.

This also fixes the duplicate notifications bug where we send as many
notifications as destinations we have in case they use the same internal
topic.

Issue: BB-507