Handle retained buckets in connectors and immutable pipelines
When removing a connector from a pipeline, the bucket is not
immediately removed from it: we udpate the in-memory map and update
the pipeline during the next update from the connectorsManager.
This may take up to several seconds. Meanwhile, we might have an
event where the same bucket is recreated, and assign it to another
connector. As a result, we might have trwo connectors handling
the same bucket for a small amount of time, with risks of having
duplicates.

This PR introduces a new method for the allocation strategy, i.e.
a check to know if the pipeline can be updated.
In addition, the ImmutableStrategy is added, to handle immutable
pipelines, where it is not possible to add or remove buckets.

Finally, the RetainedBucketsDecorator extends the AllocationStrategies
by providing support for retained buckets, for all allocation
strategies. It exposes two additional functions (as callback) to
maintain a valid state. When a retained bucket is recreated, the
connector where the bucket was before (still in use) is kept, ensuring
no risk of assigning it to another connector.

Issue: BB-601