fix ambiguous redis connection config
Redis sentinel is what's deployed in S3C, to connect to it
we use "sentinels" field in redis config that contains the host/port
of each sentinel.

In Artesca we use the "host" and "port" fields to connect to the redis
instance.

The config rules were adapted to make sure we don't pass all the fields at once
which might make the config ambiguous to understand, although the redis client
seems to handle this case well, as it first checks the sentinel config before connecting
using the host/port fields.

Default values for redis host/port were removed from the config definition and put inside
the default/dev config file.

Issue: BB-522