Make allSettled usage safer
allSettled does not follow the usuage fullfil pattern: it will never
reject, and always fullfil with an array of the results of each
promises.

This is not an issue in the case of lifecycle, where we actually ignore
all errors; but it makes the code look inconsistent, as it suggests
errors are possible but not handle them.

To avoid future issues, add proper processing of the results of
allSettled to build a single error when appropriate.

Issue: BB-641