use proper label in queue populator metrics
Queue Populator can be used for a single or multiple
extensions at once. Currently the 'origin' label in the
metrics is always set to "replication" which is not always
true,  we change this to make it hold the list of loaded
extensions in the current queue populator instance.

Issue: BB-527