BB-676: Ignore APPEND_TO_LOG operations
In the context of handling bucket migration
in SUR, we have introduced a new type of
operations that appear in the Metadata oplog
and are consumed by Backbeat.

These are “appended” operations, they appear
in the oplog in order to pass information to oplog
consumers, but do not incur any change in the database state.

The format the new  operation types is (using an example):
    {
      "db": "some-bucket",
      "method": 10,
      "entries": [
        {
          "type": "bucket_migration_stage",
          "value": "{ ... }"
        }
      ],
      "timestamp": "2025-04-03T06:36:50.401Z"
    },

method: 10 is a new method type, APPEND_TO_LOG.
Only these new operation have this method type.

This change ensures that Backbeat ignores all
operations with method == 10.