BB-676: Add check for known-supported operations
In the context of handling bucket migration
in SUR, we have introduced a new type of
operations that appear in the Metadata oplog
and are consumed by Backbeat.

These are “appended” operations, they appear
in the oplog in order to pass information to oplog
consumers, but do not incur any change in the database state.

The format of new operation type is (using an example):
    {
      "db": "some-bucket",
      "method": 10,
      "entries": [
        {
          "type": "bucket_migration_stage",
          "value": "{ ... }"
        }
      ],
      "timestamp": "2025-04-03T06:36:50.401Z"
    },

method: 10 is a new method type, APPEND_TO_LOG.

To make sure that Backbeat ignores these operations
and any operations added in the future
(until we excplictly want to support new operations),
this change adds a check for operations that are known
and supported by Backbeat.