support multiple crr location in the queueProcessor
QueueProcessor currently sets the destination host to
the host of the last crr site in the bootstrap list.
This works in the 7.x branch as we filter the bootstrap
list to only contain the relevant site. This filtering
was done in the entrypoint of the QueueProcessor (i.e
task.js) which changed in 8.x and 9.x.

Issue: BB-662