Make sure calls to consume() are sequential
Calls to consume() should be done sequentially to make sure messages
are consumed in order, as there might be delay in the transport
layer causing the responses of the consume() calls to be received
out of order in a case of concurrent calls.

Consuming the messages out of order also triggers a behaviour in the
OffsetLedger where it considers this a replay scenario (can happen
after a rebalance) and proceeds by removing all offsets larger than
the most recently consumed message, this causes the offsets to never
be committed until new messages with a larger offset are consumed.
In Zenko this triggers lag alerts which is problematic.

Issue: BB-666