use auth v4 in backbeat client
Modify the backbeat client definition file to force it to use authv4,
so that Cloudserver contacts Vault with POST requests, avoiding errors
due to limitations on token header size.

I had to disable signing of the POST body, which is not supported in
streaming mode by the AWS v2 client.

Issue: BB-682