logReader: manually commit offset when using kafka logConsumer
Kafka log consumer currently auto-commits messages after they
are consumed. This is an issue as it means we loose messages
after restart of crash even if their processing didn't finish.

Like we do with other logConsumer implementations, we now manually
store offsets after the end of each batch. The auto-commit mechanism
commits offsets stored locally each interval.

Contrary to the BackbeatConsumer, there is no risk of offsets being
committed in disorder as only one batch is processed at a time, and
a partition is only assigned to a single instance of the QueuePopulator.

Issue: BB-698