Prevent crash when processing a paused location
A race condition could occur in `IngestionPopulator` if a location
is paused while its `logReader` is being processed. This pause action
removes the `logReader` from the active list.

If the asynchronous processing completes and attempts a recursive call
to `_processLogReaderEntries`, it could do so with a `logReader` that
is now `undefined`, causing a crash.

Issue: BB-716