Fix transitionTime for ExpiredObjectDeleteMarker
EODM happens when a DeleteMarker is the only remaining revision of an
object.

When expiring it, either because `ExpiredObjectDeleteMarker` is set in
the rule or because it matches the age criteria, the `transitionTime`
is thus not intrinsic to the DeleteMarker but really depends on the
time the other revisions were removed.

We don't have this information, so being precise is not possible here;
however we must not use the DeleteMarker's `LastModified` date:
- it creates seemingly very slow expiration latency metrics (since the
  DeleteMarker may be arbitrary old, even though expiration is possible
  only for a very short time)
- which triggers alerts about slow lifecycle, even though it is working
  as expected.

In addition, depending on the exact rule, there may not be an age
limit in that specific rule, which causes `getTransitionTimestamp()` to
return `null`, which prevents collecting the metrics properly.

Consequently, we now just use the current time in that case, to avoid
making metrics suspicious and triggering alert.

Issue: BB-719